Rebase main and update singleton, add class object, change float to double

+++ Ellipse/Ellipse.cpp
@@ -1,7 +1,7 @@
 #include "pch.h"
 #include "Ellipse.h"
 
-myEllipse::Ellipse::Ellipse(float semi_minor_axis, float semi_major_axis)
+myEllipse::Ellipse::Ellipse(double semi_minor_axis, double semi_major_axis)
 {
     if (semi_minor_axis < 0) {
         throw exception("The semi-minor axis of a ellipse cannot be negative or zero.");
@@ -15,14 +15,14 @@ myEllipse::Ellipse::Ellipse(float semi_minor_axis, float semi_major_axis)
     }
 }
 
-float myEllipse::Ellipse::area()
+double myEllipse::Ellipse::area()
 {
     return (PI*_semi_minor_axis * _semi_major_axis);
 }
 
-float myEllipse::Ellipse::perimeter()
+double myEllipse::Ellipse::perimeter()
 {
-    float h = pow(_semi_minor_axis - _semi_major_axis, 2) / pow(_semi_minor_axis + _semi_major_axis, 2);
+    double h = pow(_semi_minor_axis - _semi_major_axis, 2) / pow(_semi_minor_axis + _semi_major_axis, 2);
 
     return (PI*(_semi_minor_axis + _semi_major_axis)*(1 + ((3*h)/(10 + sqrt(4 - 3*h)))));
 }
@@ -32,12 +32,12 @@ string myEllipse::Ellipse::toString()
     return "Ellipse";
 }
 
-float myEllipse::Ellipse::semi_major_axis()
+double myEllipse::Ellipse::semi_major_axis()
 {
     return (_semi_major_axis);
 }
 
-float myEllipse::Ellipse::semi_minor_axis()
+double myEllipse::Ellipse::semi_minor_axis()
 {
     return (_semi_minor_axis);
 }
